publ-ish-backend-v2/requirements.txt
# --- Web Framework ---
Flask==3.0.0
Flask-Cors==4.0.0

# --- External Communication ---
requests==2.31.0

# --- NLP & Keyword Extraction ---
yake==0.4.8

# --- Semantic Search & Machine Learning ---
sentence-transformers>=3.0.0
torch>=2.6.0
numpy>=2.0.0

# --- Generative AI (RAG) ---
google-generativeai==0.3.2

# --- Utilities ---
python-dotenv==1.0.0